完成项目提交审核申请接口开发

+++ scw-project/src/main/java/com/dragon/scw/project/vo/req/ProjectBaseInfoVo.java
@@ -33,7 +33,7 @@ public class ProjectBaseInfoVo extends BaseVo {
 	private String remark;// 项目简介
 	 
 	@ApiModelProperty("筹资金额")
-	private Integer money;// 筹资金额
+	private Long money;// 筹资金额
 	 
 	@ApiModelProperty("筹资天数")
 	private Integer day;// 筹资天数
